+++ app/views/layouts/application.html.slim
@@ -9,6 +9,7 @@ html
 
     = javascript_importmap_tags
   body
-    = render "layouts/flash"
+    = turbo_frame_tag "flash"
+      = render "layouts/flash"
     main
       = yield
