+++ app/views/turbo/teachers/quizzes/new.turbo_stream.slim
@@ -1,0 +1,1 @@
+= turbo_stream.update "flash", partial: "layouts/flash"
